requirements.txt
Pillow==9.0.1
pygame==2.1.2
